refactor: 도메인 이벤트/애플리케이션 이벤트/support/인프라 레벨 등으로 리팩토링
기존에는 다음과 같이 혼재 되었음

```md
application/
├── coupon
│   └── evnet/ CouponUseEventListener.java
└── like/
│   └── event/ ProductLikeEvent.java , LikeEventListener.java
└── payment
│   └── event/ PgPaymentRequestedEvent.java, PgPaymentEventListener.java

domain/
├── tracking/
│   └── general 공통 유저 행동 이벤트 모델을 정의 / 커스텀 애노테이션 구현
│   └── order 주문 이벤트 관련 Payload, 이벤트 모델, 커스텀 애노테이션 구현

interfaces/
├── tracking
│   └── outbox/ OrderEventOutboxListner.java
│   └── UserActionTrackingAspect.java, OrderEventPublishingAspsect.java

```

변경 후

```md
com.loopers.domain
└── order
│   ├── OrderService.java
│   └── event
│       ├── OrderEventPayload.java
│       └── OrderEventType.java
└── like
    └── event
        └── LikeEventType.java

com.loopers.application
├── order
│   ├── UserOrderProductFacade.java
│   └── listener
│       └── OrderEventOutboxListener.java
├── like
│   ├── UserLikeProductFacade.java
│   ├── event
│   │   └── ProductLikeEvent.java
│   └── listener
│       └── LikeEventListener.java
├── coupon
│   └── listener
│       └── CouponUseEventListener.java
└── payment
    └── listener
        └── PgPaymentEventListener.java

com.loopers.infrastructure.outbox
├── OrderEventOutboxEntity.java
└── OrderEventOutboxRepository.java

com.loopers.support.tracking
├── annotation
│   ├── PublishOrderEvent.java
│   ├── TrackUserAction.java
│   └── EventActorType.java
├── general
│   ├── UserActionEvent.java
│   └── UserActionType.java
└── aspect
    ├── OrderEventPublishingAspect.java
    └── UserActionTrackingAspect.java
```
