+++ apps/commerce-api/src/main/java/com/loopers/application/order/UserOrderProductFacade.java
@@ -11,8 +11,8 @@ import com.loopers.domain.order.OrderModel;
 import com.loopers.domain.order.OrderService;
 import com.loopers.domain.product.ProductModel;
 import com.loopers.domain.product.ProductService;
-import com.loopers.domain.tracking.order.OrderEventType;
-import com.loopers.domain.tracking.order.PublishOrderEvent;
+import com.loopers.domain.order.event.OrderEventType;
+import com.loopers.support.tracking.annotation.PublishOrderEvent;
 import com.loopers.domain.user.UserModel;
 import com.loopers.domain.user.UserService;
 import com.loopers.support.error.CoreException;
