+++ apps/commerce-api/src/main/java/com/loopers/domain/order/event/OrderEventPayload.java
@@ -1,4 +1,4 @@
-package com.loopers.domain.tracking.order;
+package com.loopers.domain.order.event;
 
 import com.loopers.domain.order.OrderStatus;
