+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/product/ProductV1Controller.java
@@ -3,9 +3,9 @@ package com.loopers.interfaces.api.product;
 import com.loopers.application.product.ProductLikeSummary;
 import com.loopers.application.product.ProductQueryService;
 import com.loopers.domain.product.ProductSortType;
-import com.loopers.domain.tracking.general.UserActionType;
+import com.loopers.support.tracking.general.UserActionType;
 import com.loopers.interfaces.api.ApiResponse;
-import com.loopers.domain.tracking.general.TrackUserAction;
+import com.loopers.support.tracking.annotation.TrackUserAction;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
