+++ apps/commerce-api/src/main/java/com/loopers/support/tracking/general/EventActorType.java
@@ -1,4 +1,4 @@
-package com.loopers.domain.tracking.general;
+package com.loopers.support.tracking.general;
 
 /**
  * 내부 시스템에서 발생한 이벤트에 대한 액터 타입을 정의합니다.
