+++ apps/commerce-api/src/main/java/com/loopers/support/tracking/general/UserActionType.java
@@ -1,4 +1,4 @@
-package com.loopers.domain.tracking.general;
+package com.loopers.support.tracking.general;
 
 /**
  * 공통 유저 행동 이벤트 모델을 정의합니다.
